SwitchCase(Falling Through)
Write a program to know you age group.
with the help of switch case (I will try to use to fall through under
switch concept here for this.

Infants and Toddlers: 0 to 2 years old
Preschoolers: 3 to 5 years old
Children: 6 to 12 years old
Adolescents: 13 to 19 years old
Young Adults: 20 to 39 years old
Middle-Aged Adults: 40 to 59 years old
Older Adults: 60 to 79 years old
Elderly or Senior Citizens: 80 years old and above

switch case : If several options are available it is not recommended to
used nested if-else  because it reduces the readability of the code. To
handle this we should go for switch statement.

RULE TO FOLLOW:

1. IT SHOULD BE CONSTAND EXPRESSION.
2. THE VALUE SHOULD BE IN THE RANGE OF SWITCH ARGUMENT TYPE.
3. DUPLICATES CASE LABLE ARE NOT ALLOWED.